Fix for issue 1370 : tests for resynchronization after import or restore are failing This was happening because in some cases the Changelog server was failing to re-send the changes to the LDAP servers.
A new test have been added to test the Changelog server in this case.

It was also necessary to add some timeout in the test setup to let some time to
the synchornization to setup before starting the actual test. The tests have
therefore been marked as slow.

This commit also enables a Schema Synchronization test that was left disabled by error
in a previous commit



git-svn-id: https://svn.forgerock.org/opendj/trunk@1493 41b1ffd8-f28e-4786-ab96-9950f0a78031