Fix for issue 1432 (Uninstall fails on windows due to lock on activation.jar).
There are two things fixed here: with AspectJ we have more jar dependencies when running the uninstaller so just tell the user at the end to delete the files under the lib directory (instead of just deleting a couple of jars).

The other problem that I have encountered is that when we try to delete the server.lock file (even if the server is not running) we cannot do it because there are some references to the file.  This is caused because when we call the method LockFileManager.acquireExclusiveLock when the acquisition of the lock fails we do not close the FileChannel nor the RandomAccessFile that we are opening to try to get the lock.  The modifications in LockFileManager just close the RandomAccessFile and the FileChannel that have been opened when we could not get the lock (is just a clean up).


git-svn-id: https://svn.forgerock.org/opendj/trunk@1514 41b1ffd8-f28e-4786-ab96-9950f0a78031