The following changes have two main goals:
1. Fix a bug in the status panel code.  The status panel code assumes that there is only one base DN per backend and is not able to display several base DNs per database.

2. Extend the information displayed by the status panel in order to show synchronization status.  Basically we say if a suffix is synchronized or not and the number of missing changes and the age of the oldest missing change for the synchronized suffixes.

Please note that due to the fact that there is quite a lot of information to be displayed, the output of the status command line has changed.  Instead of displaying a table-like layout for the  Data Sources, we display the information for a given base DN in a block. This is done mainly because displaying a table with 6 columns in lines of 80 characters is complicated if we do not want to use abreviations.



git-svn-id: https://svn.forgerock.org/opendj/trunk@1651 41b1ffd8-f28e-4786-ab96-9950f0a78031