This change (no associated issue no.) addresses some counter-intuitive behavior discovered when specifying some duration based properties:
   * duration and size units had to be specified in default behaviors

   * upper/lower limit had to be specified without a unit and with
     values in the property's base unit (bytes for size properties)

This change enables us to specify duration and size property constraints and defaults with or without units (when no unit is specified we default to the property's base unit).

The Duration{Unit|PropertyDefinition} APIs have also been re-aligned with the Size{Unit|PropertyDefinition} APIs where possible and the unit tests updated (incl. replacing many bad uses of the assert keyword with the assertEquals TestNG method).

git-svn-id: https://svn.forgerock.org/opendj/trunk@1689 41b1ffd8-f28e-4786-ab96-9950f0a78031