This is is the last part of the replication renaming (issue 1090) and is the renaming of the configuration.
The change in the configurations are quite simples, basically the changelog string
has been replaced by the replication-server and the synchronization string
has been replaced by replication except when it is used as synchronization provider. 

The schema is a bit more complex because I've decided to keep the old objectclass
names and the old attribute names as alias of the new names.
This would allow the previous configuration to work, however my testing has shown that
the admin framework does not seem to deal well with objectclass aliases and therefore
this is not entirely true.


git-svn-id: https://svn.forgerock.org/opendj/trunk@1725 41b1ffd8-f28e-4786-ab96-9950f0a78031