Add initial support for server-side sorting in OpenDS.  This implementation will only work for indexed searches, and it operates by sorting the ID list before iterating through the entries to return them to the client.
The ldapsearch tool has also been updated to support the "-S" option to
specify that the entries should be sorted with a given sort order.

A more complete sorting solution, which will offer the ability to sort results
that are not otherwise indexed, will require a sort/VLV index mechanism like
that described in issue #38.

OpenDS Issue Number:  79


git-svn-id: https://svn.forgerock.org/opendj/trunk@1732 41b1ffd8-f28e-4786-ab96-9950f0a78031