Issue [1024] Invalid filters are not detected by ldapsearch
This change addresses a specific case of the problem identified in the issue. Specifically, a test was added to detect and reject the specific case of apostrophes (single-quotes) enclosing an LDAP filter, which occurs when ldapsearch is invoked on a Microsoft Windows system and the search filter has mistakenly been enclosed by apostrophes instead of quotation marks.

Issue 1565 has been opened to address the more general question of what constitutes a valid search filter, including the characters allowed in an attribute type.

Thanks to Andy Coulbeck for help analyzing the problem and designing a solution, and David Ely for reviewing the proposed change.


git-svn-id: https://svn.forgerock.org/opendj/trunk@1774 41b1ffd8-f28e-4786-ab96-9950f0a78031