This checkin includes a few changes, but the only real problem that I saw was here:
-    debug("Deleting the service '%s'.");
+    debug("Deleting the service '%s'.", serviceName);

Which was causing the process to core on uninstall.  So I went through the other debug messages, and made a few more changes that most likely wouldn't have caused problems.  For instance,

-      debug("doStartApplication: spawn failed.  Sent command:");
-      debug(command);
+      debug("doStartApplication: spawn failed.  Sent command: '%s'", command);

Technically, command could include a format directive (e.g. %s) in which case we'd have the same problem as debug("Deleting the service '%s'."), but this is unlikely.  There are also a few other debug messages that I added to track down the real problem, and I've left them in.


git-svn-id: https://svn.forgerock.org/opendj/trunk@1833 41b1ffd8-f28e-4786-ab96-9950f0a78031