Contribution commit from Chris Drake for issue 1103 : Add monitor data for naming/modify conflicts This makes available 3 new counters in the replication monitoring information providing :  - the number of unresolved naming conflicts since last startup : unresolved-naming-conflicts  - the number of resolved naming conflicts since last startup : resolved-naming-conflicts  - the number of resolved modify conflicts since ast startup : resolved-modify-conflicts
Note that modify conflicts are always resolved automatically.
Also note that these counters are reset when the server is restarted.

This commit also include a new test for modify conflicts with some checks of the counters values
and update to the naming conflict test in order to test the counter implementation.

Thanks again to Chris

git-svn-id: https://svn.forgerock.org/opendj/trunk@1966 41b1ffd8-f28e-4786-ab96-9950f0a78031