Fix for 1323 : Error message on startup with synchronization enabled
When 2 replication server instances open the connection to each other simultaneously
one of this connection must be closed. In that case an error message 
was produced even though this is not an error condition.

This change removes the error message in such condition and also improves the
message that get printer to the error log in the case where 2 different server
have been assigned with the same server ID.

git-svn-id: https://svn.forgerock.org/opendj/trunk@1968 41b1ffd8-f28e-4786-ab96-9950f0a78031