This commit essentially removes the step of backing up the databases as part of an upgrade.  There are concerns that this may take a long time or eat up lots of resources in cases where there is alot of data.  Also if the data is important it is likely that there will be a separate backup strategy for this.
Since it might be nice to offer to do this as an optional step in the future, I have not removed the code altogether but rather surrounded the backup invocation code with a conditional that is always false right now.  When we can resolve where in the GUI to offer such a checkbox for managing the value of this value there will be no change necessary to the Upgrader class. 


git-svn-id: https://svn.forgerock.org/opendj/trunk@1985 41b1ffd8-f28e-4786-ab96-9950f0a78031