Update the server to provide a lockdown mode.  This is a mode in which the server will only allow client connections over loopback interfaces and will reject requests from non-root users.  This can be used in cases where it would be helpful for the server to be online to address a problem, but there might be security risks in having it fully available (e.g., the server detects a malformed access control rule on startup, and we don't want to allow normal access to the server since that rule might be intended to prevent users from seeing sensitive information and not having it interpreted properly could be dangerous).
This mode is designed so that server components like the access control
subsystem can place the server in this mode if a problem is detected, but it
also includes tasks that can be used to manually place the server into and out
of the lockdown mode.  These tasks will only be allowed to be invoked by root
users over a loopback connection.

OpenDS Issue Number:  1758


git-svn-id: https://svn.forgerock.org/opendj/trunk@2002 41b1ffd8-f28e-4786-ab96-9950f0a78031