Fix for issue 1761.
The problem was that the code of the WebStartInstaller tried to create an Installation object.  Installation calls start-ds(.bat) to get information about the build, but this fails if no JAVA_HOME is set.  JAVA_HOME is not set in general for the WebStart installer but it should be for the other applications.  The fix consists of writing the java home before creating the Installation object.

Some fixes have been also done related to the progress bar of the web start installer.  Now there is some progress bar updates during the downloading.


git-svn-id: https://svn.forgerock.org/opendj/trunk@2009 41b1ffd8-f28e-4786-ab96-9950f0a78031