Fix for issue 1601.
Basically the fix consists of making the progress panel JEditorPane not to be the root of the focus cycle.  Appart from that some focus listeners have been added to be sure that the first component that gets the focus is the details text area in the progress panel.

Concerning the status panel, the panel could be used using just the keyboard, but when the tables get the focus CTRL+TAB must be used.  As the tables are read-only there is no much value on making them focusable so I set them to be non focusable so that the use of CTRL+TAB is no longer required to use the status panel.


git-svn-id: https://svn.forgerock.org/opendj/trunk@2017 41b1ffd8-f28e-4786-ab96-9950f0a78031