Provide a new mechanism for encoding entries.  This method adds an extra element that includes flags that indicate how the entry was encoded.  The flags currently defined include:
- Whether to exclude the DN from the encoded entry.  This may be useful for the
  filesystem entry cache, since it will already have a reference to the DN.

- Whether to compress the set of object classes contained in the entry with a
  binary token.

- Whether to compress the attribute descriptions in the entry with binary
  tokens.

These changes can help improve the encode/decode performance, and can reduce
the entry footprint (by about 30% for entries based on the example.template).
None of these options are enabled at the present time, but components which
call the Entry.encode() method will be able to indicate which of them should be
used for that entry.

OpenDS Issue Numbers:  660, 1675, 1770


git-svn-id: https://svn.forgerock.org/opendj/trunk@2049 41b1ffd8-f28e-4786-ab96-9950f0a78031