Commit quite a lot of bug fixes to the ADS.
The fixes included the handling of quite a lot of errors that might be generated by the Administration Framework.

The creation of the Administrators has been revisited and now they use the privilege system and no acis are defined (which implies that no modification to the configuration is required in order the administrators to work).


git-svn-id: https://svn.forgerock.org/opendj/trunk@2053 41b1ffd8-f28e-4786-ab96-9950f0a78031