The following commit adds all the code necessary to be able to configure replication using the setup.
Some tests have been done to check that the code is working but there are some additional things to verify before making the new functionalities visible.  That is why the new functionalities have not been made available in the setup interface.

In order to be able to see the new functionalities some commented lines must be uncommented in Uninstaller.java (search for some 'TODO' comments in the code).


git-svn-id: https://svn.forgerock.org/opendj/trunk@2054 41b1ffd8-f28e-4786-ab96-9950f0a78031