Fix for 1797 : dead locks during replication testing
A listener thread has received an update message, processed it and during the
postOperation phase is now trying to send update that were in the PendingChanges Queue to the Replication Server.
However it is blocked because the protocol window is closed.
Unfortunately it is still holding the PeningChanges locks and therefore
the other listenerthread are blocked and don't read the ack messages that would open the window.

The solution is to split the pendingChanges is 2 queues : one for the local update and one for the remote updates so that the receiving threads are not locked anymore because the window is closed.

git-svn-id: https://svn.forgerock.org/opendj/trunk@2118 41b1ffd8-f28e-4786-ab96-9950f0a78031