Add a new extended operation that can be used to interact with password policy state information for a user, including getting and setting various state variables (even those marked NO-USER-MODIFICATION, although these are typically intended for testing purposes only).  Users will be required to have the password-reset privilege in order to be able to use this extended operation, and access control will also come into play.
Also, include a manage-account tool that can be used to interact with this
extended operation.  It supports a subset of the capabilities made available in
the extended operation (e.g., it only allows one operation per use, whereas the
extended operation allows you to include multiple operations), and some of the
features intended for testing purposes are marked hidden so that they aren't
readily apparent to end users.

Finally, fix a bug in which it was not possible for users to authenticate if
their account had the pwdReset flag and last login time tracking was enabled.

OpenDS Issue Numbers:  292, 579, 1782, 1845


git-svn-id: https://svn.forgerock.org/opendj/trunk@2201 41b1ffd8-f28e-4786-ab96-9950f0a78031