This commit's design changes have been reviewed and approved by Brian Ehret.
Add a "Finish" step to the graphical upgrade, setup and uninstall.

Issue 1325: Update quicksetup to integrate common install scenario

These modifications allow to configure replication in the graphical setup.

Issue 1850: consider replication for uninstall and cancel install

These modfications allow to cancel the modifications done in remote servers during the install when the user clicks on cancel.

Issue 1863: combine Web Start upgrade and install into one tool

These modifications allow to launch an application that combines both install and upgrade.


Note: because of some problems with the total update (see Issue 1893) the administration suffix is configured with its own database so a new directory has been created in the installation (adminDb).  If we consider that is not a good idea to have a separate database this will be updated in the future once 1893 is fixed.  The directory must be created during installation because in some case the Administration configuration is created using LDAP and remotely when installing a new server (and in this case we have no access to the fiel system of the firstly installed server).


git-svn-id: https://svn.forgerock.org/opendj/trunk@2257 41b1ffd8-f28e-4786-ab96-9950f0a78031