fix for 1898 : Conflicts between moddn and del generate server errors
In the solveNamingConflict method, the case when the entry has been
deleted was not correctly tested, thus leading to a NullPointerException,
The fix is simply to return since when the entry has already been deleted
there is not need to change its name anymore.

A test for this case has also been added.

git-svn-id: https://svn.forgerock.org/opendj/trunk@2280 41b1ffd8-f28e-4786-ab96-9950f0a78031