Fix issue 1826: delete of missing managed objects should throw error in dsconfig
This change fixes a bug in the admin framework where an attempt to delete a non-existent managed object would not throw a ManagedObjectNotFoundException. In addition, a "-f" (or "--force") option has been added to the dsconfig delete-xxx sub-command to simulate the "-f" usage supported in the Unix "rm" command. If the "-f" option is specified, then no error will be returned to the client if the named managed object or its parents do not exist.


git-svn-id: https://svn.forgerock.org/opendj/trunk@2286 41b1ffd8-f28e-4786-ab96-9950f0a78031