Issue 874 ldifsearch always returns objectclass attribute.
Also fixed the -o outputfile option, which was ignoring the specified filename and
incorrectly wrote the output to the current working directory using the default
filename (output). 

Lastly, added a boolean param to mainSearch (similar to the other tools), selecting if
server initialization should be performed or not. This was needed so the unit test would 
run properly.


git-svn-id: https://svn.forgerock.org/opendj/trunk@2289 41b1ffd8-f28e-4786-ab96-9950f0a78031